- Add order by in querry - Add session class to get user id - Add flash messages - Add authorisation (if user owns a resource)

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController {
 
@@ -22,7 +24,7 @@ class ListingController {
 
     public function index(){
 
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', [
             'listings'=> $listings
@@ -79,7 +81,7 @@ class ListingController {
 
     $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
 
     $newListingData = array_map('sanitize', $newListingData);
 
@@ -125,6 +127,8 @@ class ListingController {
         $query = "INSERT INTO listings ({$fields}) VALUES ({$values})";
         $this->db->query($query, $newListingData);
 
+        Session::setFlashMessage('success_message', 'Listing created successfully');
+
         redirect('/listings');
     }
  }
@@ -145,16 +149,22 @@ class ListingController {
 
     $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+    //check if listing exist
     if(!$listing){
         ErrorController::notFound('Listing not found');
         return;
     }
 
+    //authorization
+    if(!Authorization::isOwner($listing->user_id)){
+        Session::setFlashMessage('error_message', 'You are not authorised to delete this listing');
+        return redirect('/listings/'.$listing->id);
+    }
+
     $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
     //set flash message
-    $_SESSION['success_message'] = 'Listing deleted successfully';
-
+    Session::setFlashMessage('success_message', 'Listing deleted successfully');
     redirect('/listings');
   }
 
@@ -179,6 +189,12 @@ class ListingController {
         if(!$listing){
             ErrorController::notFound('Listing not Found');
             return;
+        }
+
+         //authorization
+         if(!Authorization::isOwner($listing->user_id)){
+            Session::setFlashMessage('error_message', 'You are not authorised to update this listing');
+            return redirect('/listings/'.$listing->id);
         }
 
         loadView('listings/edit', [
@@ -206,6 +222,12 @@ class ListingController {
         if(!$listing){
             ErrorController::notFound('Listing not Found');
             return;
+        }
+
+         //authorization
+        if(!Authorization::isOwner($listing->user_id)){
+            Session::setFlashMessage('error_message', 'You are not authorised to update this listing');
+            return redirect('/listings/'.$listing->id);
         }
 
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
@@ -247,7 +269,7 @@ class ListingController {
                 $updateValues['id'] = $id;
                 $this->db->query($updateQuery, $updateValues);
         
-                $_SESSION['success_message'] = 'Listing updated successfully';
+                Session::setFlashMessage('success_message', 'Listing updated successfully');
         
                 redirect('/listings/' . $id);
             }
